:recycle: Refactor: BaseEntity, User 도메인 코틀린 리팩토링
