+++ kyonggiyo-core/src/main/java/kyonggiyo/domain/BaseEntity.kt
@@ -1,0 +1,23 @@
+package kyonggiyo.domain
+
+import jakarta.persistence.Column
+import jakarta.persistence.MappedSuperclass
+import jakarta.persistence.PreUpdate
+import java.time.LocalDateTime
+
+@MappedSuperclass
+abstract class BaseEntity {
+
+    @Column(name = "created_at", nullable = false, updatable = false, columnDefinition = "TIMESTAMP(6)")
+    val createdAt: LocalDateTime = LocalDateTime.now()
+
+    @Column(name = "updated_at", nullable = false, columnDefinition = "TIMESTAMP(6)")
+    var updatedAt: LocalDateTime = LocalDateTime.now()
+        protected set
+
+    @PreUpdate
+    fun preUpdate() {
+        updatedAt = LocalDateTime.now()
+    }
+
+}
