+++ kyonggiyo-core/src/main/java/kyonggiyo/domain/user/User.kt
@@ -1,38 +1,30 @@
 package kyonggiyo.domain.user
 
+import jakarta.persistence.*
+import kyonggiyo.domain.BaseEntity
 
+@Entity
+@Table(name = "users")
 class User(
-        val id: Long,
-        val role: Role,
-        val nickname: String,
-        val isDeleted: Boolean,
-) {
-    override fun equals(other: Any?): Boolean {
-        if (other is User) {
-            return this.id == other.id
-        }
-        return false
-    }
+    role: Role,
+    nickname: String
+) : BaseEntity() {
 
-    override fun hashCode(): Int {
-        return super.hashCode()
-    }
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long? = null
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    var role: Role = role
+        protected set
+
+    @Column(unique = true, nullable = false)
+    var nickname: String = nickname
+        protected set
+
+    @Column(nullable = false)
+    var isDeleted: Boolean = false
+        protected set
+
 }
-
-//@Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//
-//    @Enumerated(EnumType.STRING)
-//    private Role role;
-//
-//    @Column(unique = true)
-//    private String nickname;
-//
-//    private boolean isDeleted;
-//
-//    public User(String nickname) {
-//        this.role = Role.USER;
-//        this.nickname = nickname;
-//        this.isDeleted = false;
-//    }
